Fix tests that fail with pytest
